+++ test/system/shell/zsh/helper.py
@@ -27,7 +27,8 @@ def zsh_shell():
     """Spawns a Zsh shell and provides Envprobe in the PATH."""
     envprobe_at = envprobe_location()
 
-    with Shell("zsh", "--no-rcs --no-promptsp --interactive --shinstdin",
+    with Shell("zsh", "--no-rcs --no-globalrcs --no-promptsp "
+               "--interactive --shinstdin",
                "echo $?", ";\n", is_interactive=True) as shell:
         _, pid = shell.execute_command("echo $$")
         print("[Zsh] Spawned {0}".format(pid), file=sys.stderr)